requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
altair==5.5.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.13.1
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
grad-cam==1.5.5
idna==3.10
Jinja2==3.1.3
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
labelImg==1.8.6
lxml==6.0.2
MarkupSafe==2.1.5
matplotlib==3.10.6
mpmath==1.3.0
multidict==6.6.4
narwhals==2.6.0
networkx==3.3
numpy==2.1.2
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.0.0
playsound==1.3.0
polars==1.34.0
polars-runtime-32==1.34.0
propcache==0.4.0
protobuf==6.32.1
psutil==7.1.0
pyarrow==21.0.0
pydeck==0.9.1
PyJWT==2.10.1
pyparsing==3.2.5
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
smmap==5.0.2
streamlit==1.50.0
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.2
tqdm==4.67.1
ttach==0.0.3
twilio==9.8.3
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.204
ultralytics-thop==2.0.17
urllib3==2.5.0
watchdog==6.0.0
yarl==1.20.1
